Major changes in code to make it work for SHA  * Changes cannot be trusted, though seem to work superficially for sieve and adpcm  * Changes in the IR_C code for SHA to make it work with the tool  * Instrumented SHA does not compile yet, and fails with many error. Next Task!

+++ instrument/examples/sieve/instrumented/erat_sieve_no_print_IR.c
@@ -25,7 +25,7 @@ struct test {
   unsigned int v;
   unsigned int k;
 } m = { 1, 1 };
-unsigned long m_addr = 0x7c8;
+unsigned int m_addr = 0x7c8;
 
 void sieve_func() {
   int j_76;
